factored out is_variable_proc to a header file
